refactor(client): replace previously used filter, sort and pagination types
