--- apps/next-app/src/components/collection/filter/types/index.ts
@@ -1,10 +1,0 @@
-import type { Filter, Sort } from "@/lib/types/query"
-import type { CollectionFilterQuery, CollectionSortQuery } from "@/lib/schema/query/collection-query"
-
-/* Filter types */
-export type CollectionFilterFields = Required<CollectionFilterQuery>
-export type CollectionFilter = Filter<CollectionFilterFields>
-
-/* Sort types */
-export type CollectionSortFields = Required<CollectionSortQuery>
-export type CollectionSort = Sort<CollectionSortFields>
