+++ apps/next-app/src/server/db/query/session-query.ts
@@ -2,7 +2,7 @@
 import type { MatchFormat } from "@repo/db"
 import type { ClientSessionVariants } from "@repo/schema/session"
 import type { GameSessionWithPlayersWithAvatarWithCollectionWithCards } from "@repo/db/types"
-import type { Pagination, PaginationParams } from "@/lib/types/query"
+import type { Pagination, PaginationParams } from "@repo/schema/search"
 import type { SessionFilterQuery, SessionSortQuery } from "@/lib/schema/query/session-query"
 
 // db
